Isolate tinyhtml5 compatibility layer
